Added the quotation view with steps
