+++ web/src/hooks/api/useLeadStepStatuses.ts
@@ -4,6 +4,7 @@ import { useLeadEnquiries } from "@/hooks/api/useLeadEnquiry";
 import { useQuery } from "@tanstack/react-query";
 import { leadEnquiryService } from "@/services/api/lead-enquiry.service";
 import { enquiryCostingService } from "@/services/api/enquirycosting.service";
+import { leadsQuotationService } from "@/services/api/leads-quotation.service";
 import type { StepStatus } from "@/components/layout/ShowPageLayout";
 
 function deriveStatus(hasData: boolean, isLoading: boolean): StepStatus {
@@ -39,6 +40,12 @@ export function useLeadStepStatuses(leadId: number | null) {
     const { data: costings, isLoading: l5 } = useQuery({
         queryKey: ['enquiry-costings', 'by-lead', leadId],
         queryFn: () => enquiryCostingService.getByLeadId(leadId!),
+        enabled: !!leadId,
+    });
+
+    const { data: quotations, isLoading: l6 } = useQuery({
+        queryKey: ['leads-quotations', 'by-lead', leadId],
+        queryFn: () => leadsQuotationService.getByLeadId(leadId!),
         enabled: !!leadId,
     });
 
@@ -87,6 +94,15 @@ export function useLeadStepStatuses(leadId: number | null) {
             hasData: Array.isArray(costings) && costings.length > 0,
             isLoading: l5,
             status: deriveStatus(Array.isArray(costings) && costings.length > 0, l5),
+        },
+        {
+            id: "quotations",
+            label: "Quotations",
+            shortLabel: "Quot",
+            stepNumber: 6,
+            hasData: Array.isArray(quotations) && quotations.length > 0,
+            isLoading: l6,
+            status: deriveStatus(Array.isArray(quotations) && quotations.length > 0, l6),
         },
     ];
