+++ web/src/modules/crm/lead-enquiry/LeadEnquiryViewPage.tsx
@@ -13,6 +13,7 @@ import { LeadDetailsSection } from "../leads/components/LeadView";
 import { FollowupViewPage } from "../followups/FollowupViewPage";
 import { LeadSiteVisitsSection } from "./components/LeadSiteVisitView";
 import { LeadCostingsSection } from "../enquirycosting/EnquiryCostingViewPage";
+import { LeadQuotationsSection } from "../leads-quotation/LeadsQuotationViewPage";
 
 interface LeadEnquiryViewProps {
     enquiry?: LeadEnquiryWithNames | null;
@@ -221,6 +222,8 @@ export function LeadEnquiryViewPage({ enquiryId, onBack, backLabel }: LeadEnquir
                 return <LeadSiteVisitsSection leadId={leadId} />;
             case "costings":
                 return <LeadCostingsSection leadId={leadId} />;
+            case "quotations":
+                return <LeadQuotationsSection leadId={leadId} />;
             default:
                 return null;
         }
