+++ web/src/modules/crm/leads/LeadViewPage.tsx
@@ -6,6 +6,7 @@ import { useLeadStepStatuses } from "@/hooks/api/useLeadStepStatuses";
 import { LeadEnquiriesSection } from "../lead-enquiry/LeadEnquiryViewPage";
 import { LeadSiteVisitsSection } from "../lead-enquiry/components/LeadSiteVisitView";
 import { LeadCostingsSection } from "../enquirycosting/EnquiryCostingViewPage";
+import { LeadQuotationsSection } from "../leads-quotation/LeadsQuotationViewPage";
 
 interface LeadViewPageProps {
     leadId: number;
@@ -59,6 +60,8 @@ export function LeadViewPage({ leadId, onBack, backLabel }: LeadViewPageProps) {
                     return <LeadSiteVisitsSection leadId={leadId} />;
                 case "costings":
                     return <LeadCostingsSection leadId={leadId} />;
+                case "quotations":
+                    return <LeadQuotationsSection leadId={leadId} />;
                 default:
                     return null;
             }
